+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## [Unrelease]
 ### Added
+- Add the python code `generatePkgConfig.py` to parse argument from command line
 
 ### Fixed